+++ basic-arch-rbac/rbac-service/src/main/java/com/gzs/learn/rbac/po/RolePo.java
@@ -23,7 +23,7 @@ public class RolePo extends Base {
     /**
      * 父角色id
      */
-    private Integer pid;
+    private Long pid;
     /**
      * 角色名称
      */
@@ -31,7 +31,7 @@ public class RolePo extends Base {
     /**
      * 部门名称
      */
-    private Integer deptid;
+    private Long deptid;
     /**
      * 提示
      */
