+++ basic-arch-web/src/main/java/com/gzs/learn/web/core/shiro/ShiroUser.java
@@ -11,8 +11,8 @@ public class ShiroUser implements Serializable {
     private Long id; // 主键ID
     private String account; // 账号
     private String name; // 姓名
-    private Integer deptId; // 部门id
-    private List<Integer> roleList; // 角色集
+    private Long deptId; // 部门id
+    private List<Long> roleList; // 角色集
     private String deptName; // 部门名称
     private List<String> roleNames; // 角色名称集
 }
